Remove User on User Mapping Table Fixes #39
