+++ app/templates/base.j2
@@ -36,10 +36,10 @@
         </button>
         <div class="collapse navbar-collapse" id="navbarNav">
             <ul class="navbar-nav">
+                {% if 'logout' not in request.path %}
                 <li class="nav-item active">
                     <a class="nav-link" href="/users">Users</a>
                 </li>
-                {% if 'logout' not in request.path %}
                 <li class="nav-item active">
                     <a class="nav-link" href="/logout">Logout</a>
                 </li>
