+++ app/templates/users.j2
@@ -2,6 +2,16 @@
 
 {% block content %}
     <h2>GitHub User Mappings</h2>
+    {% if action_message is not none %}
+    <div class="alert alert-success" role="alert">
+        <p>{{ action_message }}.</p>
+    </div>
+    {% endif %}
+    {% if error_message is not none %}
+    <div class="alert alert-warning" role="alert">
+        <p>{{ error_message }}.</p>
+    </div>
+    {% endif %}
     <table id=users class="table table-striped table-bordered" style="width:100%">
         <thead>
             <tr>
@@ -9,7 +19,7 @@
                 <th scope="col">Last</th>
                 <th scope="col">Email</th>
                 <th scope="col">Username</th>
-                <th scope="col">Status</th>
+                <th scope="col">Role</th>
                 {% if owner is sameas true %}
                 <th scope="col">Action</th>
                 {% endif %}
@@ -22,16 +32,20 @@
                 <td><p class="text-capitalize">{{ user.surname }}</p></td>
                 <td>{{ user.email }}</td>
                 <td>{{ user.username }}</td>
-                <td>{{ user.status }}</td>
+                <td>{{ user.github_role }}</td>
                 {% if owner is sameas true %}
                 <td scope="col">
-                    {% if user.status == "member" %}
-                    <button type="button" class="btn btn-warning">Convert</button>
-                    <button type="button" class="btn btn-danger">Remove</button>
-                    {% else %}
-                    <button type="button" class="btn btn-danger">Remove</button>
-                    {% endif %}
-                    </td>
+                    <form action="" method="post">
+                        <input type="hidden" name="email" value="{{ user.email }}" />
+                        <input type="hidden" name="user" value="{{ user.username }}" />
+                        {% if user.github_role == "member" %}
+                        <input type="submit" id="action" name="action" value="Convert" class="btn btn-warning" />
+                        <input type="submit" id="action" name="action" value="Remove" class="btn btn-danger" />
+                        {% else %}
+                        <input type="submit" id="action" name="action" value="Remove" class="btn btn-danger" />
+                        {% endif %}
+                    </form>
+                </td>
                 {% endif %}
             </tr>
             {% endfor %}
